fill movie info, actors etc from json real data

+++ app/src/main/java/com/application/movielist/data/JsonLoad.kt
@@ -94,7 +94,7 @@ internal fun parseMovies(
             overview = jsonMovie.overview,
             poster = jsonMovie.posterPicture,
             backdrop = jsonMovie.backdropPicture,
-            ratings = (jsonMovie.ratings / 2).toInt(),
+            ratings = (jsonMovie.ratings / 2),
             numberOfRatings = jsonMovie.votesCount,
             minimumAge = if (jsonMovie.adult) 16 else 13,
             runtime = jsonMovie.runtime,
