+++ app/src/main/java/com/application/movielist/data/MovieData.kt
@@ -5,7 +5,7 @@ data class MovieData(
     val overview: String,
     val poster: String,
     val backdrop: String,
-    val ratings: Int,
+    val ratings: Float,
     val numberOfRatings: Int,
     val minimumAge: Int,
     val runtime: Int,
